+++ src/com/kaylerrenslow/armaDialogCreator/gui/main/ADCWindow.java
@@ -6,6 +6,7 @@ import com.kaylerrenslow.armaDialogCreator.data.ImagesTool;
 import com.kaylerrenslow.armaDialogCreator.data.tree.TreeStructure;
 import com.kaylerrenslow.armaDialogCreator.gui.FXUtil;
 import com.kaylerrenslow.armaDialogCreator.gui.main.popup.ConvertingImageSubscriberDialog;
+import com.kaylerrenslow.armaDialogCreator.gui.main.popup.ConvertingImageSubscriberNotificationCreator;
 import com.kaylerrenslow.armaDialogCreator.gui.uicanvas.ScreenDimension;
 import com.kaylerrenslow.armaDialogCreator.main.ArmaDialogCreator;
 import javafx.beans.value.ChangeListener;
@@ -49,6 +50,7 @@ public class ADCWindow {
 		this.primaryStage.setScene(scene);
 
 		ImagesTool.subscribeToConversion(new ConvertingImageSubscriberDialog());
+		ImagesTool.subscribeToConversion(new ConvertingImageSubscriberNotificationCreator());
 	}
 
 	public void initialize(@Nullable TreeStructure<ArmaControl> backgroundTreeStructure, @Nullable TreeStructure<ArmaControl> mainTreeStructure) {